# docs/API.yaml
openapi: 3.0.0
info:
  title: Betting Application API
  description: API for a betting system including authentication, categories, events, and bets.
  version: 1.1.0  # Updated version to reflect changes
servers:
  - url: http://localhost:8080/api
    description: Local server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserRegistrationDTO:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - username
        - email
        - password
    UserLoginDTO:
      type: object
      properties:
        loginOrEmail:
          type: string
        password:
          type: string
      required:
        - loginOrEmail
        - password
    AuthenticationResponseDTO:
      type: object
      properties:
        tokenType:
          type: string
        accessToken:
          type: string
      required:
        - tokenType
        - accessToken
    CategoryDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
    CreatedEventDTO:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        createdOutcomes:
          type: array
          items:
            $ref: '#/components/schemas/CreatedOutcomeDTO'
      required:
        - title
        - category
        - startsAt
        - endsAt
        - createdOutcomes
    CreatedOutcomeDTO:
      type: object
      properties:
        description:
          type: string
      required:
        - description
    EventFinishDTO:
      type: object
      properties:
        eventId:
          type: integer
          format: int64
        status:
          type: string
          enum: [COMPLETED, CANCELLED]
        outcomeId:
          type: integer
          format: int64
      required:
        - eventId
        - status
    BetDTO:
      type: object
      properties:
        outcomeId:
          type: integer
          format: int64
        sum:
          type: integer
          format: int64
      required:
        - outcomeId
        - sum
    PlacedBetDto:
      type: object
      properties:
        eventId:
          type: integer
          format: int64
        outcomeId:
          type: integer
          format: int64
      required:
        - eventId
        - outcomeId
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startsAt:
          type: string
          format: date-time
        endsAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [PLANNED, ONGOING, COMPLETED, CANCELLED]
        category:
          $ref: '#/components/schemas/Category'
        outcomes:
          type: array
          items:
            $ref: '#/components/schemas/Outcome'
    Outcome:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        isWinner:
          type: boolean
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
paths:
  /auth/register:
    post:
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationDTO'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponseDTO'
        '400':
          description: Validation error
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
  /auth/login:
    post:
      summary: Login a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDTO'
      responses:
        '200':
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponseDTO'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
  /admin/category/create:
    post:
      summary: Create a new category
      operationId: createCategory
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDTO'
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Validation error or category already exists
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (requires ADMIN role)
        '500':
          description: Internal server error
  /admin/events/create:
    post:
      summary: Create a new event
      operationId: createEvent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedEventDTO'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Validation error
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (requires ADMIN role)
        '500':
          description: Internal server error
  /admin/events/finish:
    post:
      summary: Finish an event
      operationId: finishEvent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventFinishDTO'
      responses:
        '200':
          description: Event finished successfully
        '400':
          description: Validation error
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (requires ADMIN role)
        '500':
          description: Internal server error
  /events/list:
    get:
      summary: List events
      operationId: listEvents
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Category not found
          content:
            text/plain:
              schema:
                type: string
        '500':
          description: Internal server error
  /events/stream:
    get:
      summary: Stream event updates
      operationId: streamEvents
      produces:
        - text/event-stream
      responses:
        '200':
          description: Event stream
          content:
            text/event-stream:
              schema:
                type: string
  /bets/place:
    post:
      summary: Place a bet
      operationId: placeBet
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BetDTO'
      responses:
        '201':
          description: Bet placed successfully
        '400':
          description: Validation error or not enough balance
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
  /bets/list:
    get:
      summary: List user's bets
      operationId: getBetsList
      security:
        - BearerAuth: []
      parameters:
        - name: showClosed
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of placed bets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PlacedBetDto'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error